re-work gf2.h

+++ trunk/linbox/linbox/field/Modular/modular.inl
@@ -125,7 +125,7 @@ namespace LinBox {
 			y %= (uint64_t) _F._modulus;
 		}
 
-		return res = y;
+		return res = (uint16_t) y;
 	}
 
 	template <class Vector1, class Vector2>
